+++ com/src/http.c
@@ -176,7 +176,7 @@ static void naos_http_send_frame(void *arg) {
   free(msg);
 }
 
-static size_t naos_http_msg_mtu() { return 4096; }
+static uint16_t naos_http_msg_mtu() { return 4096; }
 
 static bool naos_http_msg_send(const uint8_t *data, size_t len, void *ctx) {
   // prepare message
